fixed validation error on date, restructured files
